+++ application/controllers/Upload_files.php
@@ -30,14 +30,17 @@
 			}
 		}
 
+		/**
+		 * Upload het groentje naar the sercer.
+		 */
 		public function do_upload() {
 			if($this->Session) {
-				$config = [
+				$config = array(
 					'upload_path' => './assets/files/',
 					'allowed_types' => 'pdf',
-					'file_name' => 'Planning';
+					'file_name' => 'Planning',
 					'overwrite' => TRUE,
-				];
+				);
 
 				// Library word niet constructor geladen.
 				// Omdat deze config variables bevat