+++ application/models/Model_leiding.php
@@ -17,14 +17,14 @@
 
     function Leiding_insert($Mail) {
       $Values = array(
-        "username" => $this->input->post('Naam'),
-        "Mail" => $this->input->post('Mail'),
-        "GSM" => $this->input->post('GSM'),
-        "Tak" => $this->input->post('Tak'),
+        "username"   => $this->input->post('Naam'),
+        "Mail"       => $this->input->post('Mail'),
+        "GSM"        => $this->input->post('GSM'),
+        "Tak"        => $this->input->post('Tak'),
         "Admin_role" => "0",
-        "Blocked" => "0",
-				"Theme" => "0",
-        "password" => md5($Mail['Pass']),
+        "Blocked"    => "0",
+				"Theme"      => "0",
+        "password"   => md5($Mail['Pass']),
       );
 
       $this->db->insert('users', $Values);