+++ CHANGELOG.md
@@ -8,3 +8,4 @@ Changelog
 5. Added discription off the functions. Affected = Models, controllers.
 6. Add sensio analytics and gitter support chat.
 7. Set load of helpers and libraries in arrays.
+8. Logout & goto backend added in frontend navbar.